remove individual module toggle
